+++ test/components/form/form.e2e-spec.js
@@ -6,6 +6,8 @@ requireHelper('rejection');
 
 jasmine.getEnv().addReporter(browserStackErrorReporter);
 
+const spinboxId = 'test-spinbox-age2';
+
 describe('Form Tests', () => {
   if (utils.isChrome() && utils.isCI()) {
     it('Should not visual regress on compound fields', async () => {
@@ -44,4 +46,22 @@ describe('Form Tests', () => {
       expect(await browser.protractorImageComparison.checkElement(containerEl, 'form-checkbox-in-columns')).toEqual(0);
     });
   }
+});
+
+describe('Spinbox test-spinbox-compound-field tests', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/form/test-spinbox-compound-field');
+    await browser.driver
+      .wait(protractor.ExpectedConditions
+        .presenceOf(element(by.id(spinboxId))), config.waitsFor);
+  });
+
+  it('Should not have errors', async () => {
+    await utils.checkForErrors();
+  });
+
+  it('Should align both spinbox on mobile', async () => {
+    expect(await element(by.css('.spinbox-wrapper'))
+      .getCssValue('width')).toEqual('232px');
+  });
 });